EPE-389 nodeos_irreversible_mode_lr_test fails to progress
This fix handles situations where a node is in head_catchup state and
recieves the sync_last_requested_num and then does a sync_update_expected
for the next block. And the next block has already been received by the
node. In this instance, the block will not receive the next block and
will be effectively hung.